pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.188.0",
    "isodate>=0.7.2",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "lxml[html-clean]>=6.0.2",
    "newspaper4k>=0.9.4.1",
    "ollama>=0.6.1",
    "praw>=7.8.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
    "youtube-transcript-api>=1.2.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
